study: fix parameterization ordering in failure-handler
We're not sure why this works exactly, at least we can't explain
it/reproduce it properly b/c of all the layers going on here.
We *think* what's going on is the `(parameterize (with-handlers`
ordering leads to a dynamic context that looks like

  (with-handlers
   (parameterize
    (parameterize
     (with-handlers
      ...

instead of

  (with-handlers
   (parameterize
    (with-handlers
     (parameterize
      ...

and that screws up the parameterizations of sibling/side continuations
of the exception handler (eg. those returned to the web-server via
`send/back') then reparameterized when the user continues (hits a
link).